+++ pineappl_cli/tests/evolve.rs
@@ -184,7 +184,6 @@ fn lhcb_wp_7tev() {
         .args([
             "--silence-lhapdf",
             "convolute",
-            "--scales=1",
             optimized.path().to_str().unwrap(),
             "NNPDF40_nlo_as_01180",
         ])