+++ pineappl_cli/tests/main.rs
@@ -15,11 +15,11 @@ Commands:
   import     Converts APPLgrid/fastNLO/FastKernel files to PineAPPL grids
   merge      Merges one or more PineAPPL grids together
   orders     Shows the predictions for all bin for each order separately
-  pdfunc     Calculates PDF uncertainties
   plot       Creates a matplotlib script plotting the contents of the grid
   pull       Calculates the pull between two different PDF sets
   read       Read out information of a grid
   subgrids   Print information about the internal subgrid types
+  uncert     Calculates scale and PDF uncertainties
   write      Write a grid modified by various operations
 
 Options: